+++ examples/global/global.go
@@ -1,6 +1,9 @@
 package main
 
-import "github.com/renatopp/go-cli/cli"
+import (
+	"github.com/renatopp/go-cli"
+	"github.com/renatopp/go-cli/pkg"
+)
 
 func main() {
 	cli.Name("global")
@@ -8,28 +11,28 @@ func main() {
 	cli.FlagBool("verbose", "v", "enable verbose output").AsGlobal()
 	cli.FlagInt("level", "l", "set the level").AsGlobal().WithDefault(1)
 	cli.FlagString("config", "c", "path to config file").WithDefault("config.yaml")
-	cli.Cmd("cmd1", "", cmd1)
-	cli.Cmd("cmd2", "", cmd2)
+	cli.Command("cmd1", "", cmd1)
+	cli.Command("cmd2", "", cmd2)
 	cli.Parse()
-	cli.ShowHelp()
+	cli.Help()
 }
 
 func cmd1() {
 	cli.FlagString("another", "a", "another flag").AsGlobal()
-	cli.Cmd("cmd3", "", cmd3)
+	cli.Command("cmd3", "", cmd3)
 	cli.Parse()
-	cli.ShowHelp()
+	cli.Help()
 
-	verbose, _ := cli.GetFlag[*cli.GenericFlag[bool]]("verbose")
+	verbose, _ := cli.GetFlag[*pkg.GenericFlag[bool]]("verbose")
 	println("verbose:", verbose.Value())
 }
 
 func cmd2() {
 	cli.Parse()
-	cli.ShowHelp()
+	cli.Help()
 }
 
 func cmd3() {
 	cli.Parse()
-	cli.ShowHelp()
+	cli.Help()
 }