+++ examples/hidden/hidden.go
@@ -1,14 +1,14 @@
 package main
 
-import "github.com/renatopp/go-cli/cli"
+import "github.com/renatopp/go-cli"
 
 func main() {
 	cli.Name("hidden")
 	cli.Description("A CLI with a hidden subcommands, flags and positionals.")
 	cli.AutoHelp(true)
 
-	cli.Cmd("private", "an internal command", func() {}).AsHidden()
-	cli.Cmd("public", "a public command", func() {})
+	cli.Command("private", "an internal command", func() {}).AsHidden()
+	cli.Command("public", "a public command", func() {})
 
 	cli.FlagString("secret", "s", "a secret flag").AsHidden()
 	cli.FlagString("public", "p", "a public flag")
@@ -17,5 +17,5 @@ func main() {
 	cli.Pos("public", "public positional")
 	cli.Parse()
 
-	cli.ShowHelp()
+	cli.Help()
 }